python/requirements.txt
astroid==2.1.0
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-trace==0.20.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
grpcio-tools==1.18.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
opencensus==0.2.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.2.2
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
